Upgrade to Go `1.24` and fix all golangci-lint issues
The code here needs to be upgraded to the later version of go to pull
necessary dependencies. As a result, the very out of date golangci-lint
setup and config blocks the PR merges.

Fix all issues, upgrade go mod and golangci-lint config, plus the CI
workflow to be consistent with what's used in sei-chain.